Add postcss v8 as a peer dependency
PostCSS 8 is installed as a dependency of this package, but it is not included as a peer dependency. I add this version of postcss 8 accepted versions to be consistent with the package.json installed version.

This change is required because npm versions 8.6 or higher validate peer dependencies, so a simple `npm ci` that do not have the --force flag set is going to fail breaking our CI's.